backend/requirements.txt
fastapi
hypercorn[h3]
python-multipart
python-magic
